% Generated by roxygen2 (4.0.2): do not edit by hand
\name{motiflogo}
\alias{motiflogo}
\title{A new representation of DNA motif logo}
\usage{
motiflogo(pwm, xlabel = NULL, ylabel = "Information Content",
  tit = "Motif", addpoint = FALSE, scaleSize = c(5, 20))
}
\arguments{
\item{pwm}{position weight matrix}

\item{xlabel}{text of label in x-axis}

\item{ylabel}{The default is Information Content}

\item{tit}{The default is Motif}

\item{addpoint}{add point to the position of nt. The default is FALSE}

\item{scaleSize}{the scale size of nt. ref \code{\link[ggplot2]{scale_size}}}
}
\description{
A new representation of DNA motif logo via ggplot2. Besides, the xlabel could
be DNA or SNP.
}
\examples{
pfm <- matrix(data=c(5, 3, 16, 1, 0, 17, 17, 0, 0, 16, 12, 8,
6, 9, 1, 1, 18,  1, 0, 0, 18,  1, 0, 2,
2,3, 1, 0, 0, 0, 0, 1, 0, 0, 1, 2,
5, 3, 0, 16, 0, 0, 1, 17,  0, 1, 5, 6),
byrow=TRUE,nrow=4,dimnames=list(c('A','C','G','T')))

pwm <- apply(pfm, 2, function(x) x/sum(x))
motiflogo(pwm)

xlabel <- c("T\\nT", "G\\nG", "A\\nA", "T\\nT", "C\\nC", "A\\nC", "A\\nA", "A\\nA", "C\\nC","A\\nA", "A\\nA", "T\\nT")
motiflogo(pwm, xlabel=xlabel)
motiflogo(pwm, xlabel=xlabel, addpoint=TRUE)
}

